% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistFunction.R
\name{convertDist}
\alias{convertDist}
\title{Function to convert between different units of distance}
\usage{
convertDist(x, startUnit, endUnit)
}
\arguments{
\item{x}{the number to be converted}

\item{startUnit}{the starting unit that needs to be converted}

\item{endUnit}{the unit that the data has to be converted to}
}
\value{
x converted from startUnit to endUnit
}
\description{
The function will convert from a one unit of distance to another,
with both the starting unit and ending unit being provided as parameters
in the function call
}
\examples{
convertDist(100,"cm","ft")

convertDist(212,"meter","yards")

convertDist(82,"miles","mm")

}
