% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{exchange_plot}
\alias{exchange_plot}
\title{Plots given exchange rate for last 180 days}
\usage{
exchange_plot(base, foreign)
}
\arguments{
\item{base}{the currency you want to convert from as Forex currency abbreviation}

\item{foreign}{the currency you want to convert to as Forex currency abbreviation}
}
\value{
nothing just displays a plot
}
\description{
Plots given exchange rate for last 180 days
}
\examples{
exchange_plot("EUR","GBP")
}
