% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{exchange_rate}
\alias{exchange_rate}
\title{Load in exchange rate for a particular date from last 180 days}
\usage{
exchange_rate(base, foreign, date = Sys.Date() - 1)
}
\arguments{
\item{base}{the currency you want to convert from as Forex currency abbreviation}

\item{foreign}{the currency you want to convert to as Forex currency abbreviation}

\item{date}{the date this of which you would like th exchange rate from}
}
\value{
the specified exchange rate
}
\description{
Load in exchange rate for a particular date from last 180 days
}
\examples{
exchange_rate("EUR","AUD" ,"2022-10-03")
exchange_rate("AUD","EUR" )
exchange_rate("GBP","AUD")

}
