% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popr.R
\name{fit_popr}
\alias{fit_popr}
\title{Fit model to population data}
\usage{
fit_popr(continent = "World", year1, year2, popr = load_popr())
}
\arguments{
\item{continent}{The area of the world that will be plotted}

\item{year1}{The initial year that the user wants to see the population
difference for}

\item{year2}{The final year that the user wants to see the population
difference for}

\item{popr}{The dataframe outputed from load_popr()}
}
\value{
A list that includes a dataframe of what needs to be plotted
along with the xlim and ylim for the plot
}
\description{
Fits a model that uses user imputed continents, and years to create a
dataframe and parameters for what needs to be plotted in plot_popr
}
\examples{
load_popr()
model1 <- fit_popr("World", "2022", "1970")
model2 <- fit_popr("Africa", "2015", "1990")
model3 <- fit_popr("Europe", "2010", "2000")

}
