% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.R
\name{load_fx}
\alias{load_fx}
\title{function to load in Forex data}
\usage{
load_fx(rates_list = c("EUR/USD", "EUR/GBP", "EUR/AUD"))
}
\arguments{
\item{rates_list}{A string or list of strings corresponding to Forex currency abbreviations}
}
\value{
data frame with historic exchange rates from last 180 days
}
\description{
function to load in Forex data
}
\examples{
dat1<- load_fx()
dat2<-load_fx("EUR/USD")
dat3<-load_fx(c("EUR/USD","EUR/GBP","EUR/AUD"))

}
