% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popr.R
\name{load_popr}
\alias{load_popr}
\title{Load in historical population data from Kaggle}
\usage{
load_popr()
}
\value{
A dataframe that includes population data for every country, for all 8 years.
It also includes geometry details about each of the countries.
}
\description{
Loads in population data for countries from 8 years between 1970 and 2022
}
\examples{
load_popr()

}
