% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popr.R
\name{plot_popr}
\alias{plot_popr}
\title{Plot the popr output}
\usage{
plot_popr(model = model)
}
\arguments{
\item{model}{The list outputted by fit_popr()}
}
\value{
A nice plot of the desired continent/s, with a nice colour scale.
}
\description{
Plots the users desired continent/s with the differences in populations
between the chosen years. This is accompanied with a nice colour scale
}
\examples{
load_popr()
model1 <- fit_popr("World", "2022", "1970")
model2 <- fit_popr("Africa", "2015", "1990")
model3 <- fit_popr("Europe", "2010", "2000")
plot_popr(model1)
plot_popr(model2)
plot_popr(model3)
}
